Extract pull state to its own hook to re-use it for push
